+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Restore health and armor on join
 - Copy VIN to clipboard when reading it
 - Je kan nu minimap zien in noclip
+- Alle restaurant medewerkers in dienst krijgen een percentage van het ticket als iemand anders een order maakt
 
 ### Changed